+++ src/Spinorama/index.tsx
@@ -17,6 +17,7 @@ const Spinorama: React.FC<SpinoramaProps> = (props: SpinoramaProps) => {
 
 	// Variables
 	const container = React.useRef<HTMLDivElement>(null)
+	const [selected, setSelected] = React.useState<number>(0)
 	const [totalItems, setTotalItems] = React.useState<number>(0)
 
 	// Settings
@@ -27,10 +28,21 @@ const Spinorama: React.FC<SpinoramaProps> = (props: SpinoramaProps) => {
 	}
 
 	// Callbacks
+	const nextItem = React.useCallback(() => {
+		// Find next index
+		const nextIndex = (selected + 1) % totalItems
+		setSelected(nextIndex)
+
+		return nextIndex
+	}, [selected, totalItems])
+
 	const { contextSafe } = useGSAP(
 		() => {
 			// Set interval
-			const interval = setInterval(nextItem, settings.duration)
+			const interval = setInterval(() => {
+				// Animate items
+				animateItems(nextItem())
+			}, settings.duration)
 
 			// Find total items
 			if (container.current && !totalItems) {
@@ -44,41 +56,18 @@ const Spinorama: React.FC<SpinoramaProps> = (props: SpinoramaProps) => {
 				clearInterval(interval)
 			}
 		},
-		{ scope: container, dependencies: [container, totalItems] }
+		{ scope: container, dependencies: [container, selected, totalItems] }
 	)
 
-	const findSelected = (containerElm: HTMLDivElement) => {
-		const items = containerElm.querySelectorAll('.spinorama-item')
-
-		if (items.length) return Array.from(items).findIndex(item => item.classList.contains('selected'))
-
-		return 0
-	}
-
-	const setClassName = (containerElm: HTMLDivElement, className: string, nextIndex: number) => {
-		const items = containerElm.querySelectorAll('.spinorama-item')
-
-		if (items.length)
-			items.forEach((item, index) => {
-				if (index === nextIndex) item.classList.add(className)
-				else item.classList.remove(className)
-			})
-	}
-
-	const nextItem = contextSafe(() => {
+	const animateItems = contextSafe((showIndex: number) => {
 		if (container.current && totalItems) {
-			// Find selected item
-			const selected = findSelected(container.current)
-			const nextIndex = (selected + 1) % totalItems
-			setClassName(container.current, 'selected', nextIndex)
-
 			// Find items
 			const items = container.current.querySelectorAll('.spinorama-items')
 
 			// Check items and animate it
 			if (items.length) {
 				gsap.to(items, {
-					xPercent: 100 * nextIndex,
+					xPercent: 100 * showIndex,
 					duration: settings.animateDuration,
 					ease: settings.ease
 				})
@@ -88,7 +77,12 @@ const Spinorama: React.FC<SpinoramaProps> = (props: SpinoramaProps) => {
 
 	return (
 		<Box ref={container} {...props} className={`spinorama${className ? ` ${className}` : ''}`}>
-			{children}
+			{React.Children.map(children, child => {
+				if (React.isValidElement(child))
+					// Clone element
+					return React.cloneElement(child as React.ReactElement<any>, { selected })
+				else return child
+			})}
 		</Box>
 	)
 }