+++ src/Wrapper/index.types.ts
@@ -19,4 +19,11 @@ export type SpinoramaWrapperProps = BoxProps & {
 	 * @default false
 	 */
 	reverse?: boolean
+
+	/**
+	 * Determines the currently selected slide
+	 * A number representing the index of the selected slide
+	 * If not provided, the first slide will be selected
+	 */
+	selected?: number
 }